+++ src/ft_greater.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-int		ft_greater(int num1, int num2)
+long long	ft_greater(long long num1, long long num2)
 {
 	return ((num1 > num2) ? num1 : num2);
 }
