+++ src/ft_putint.c
@@ -16,7 +16,7 @@
 ** Prints a given integer, in base-10 (decimal), to stdout.
 */
 
-void	ft_putint(int integer)
+void	ft_putint(long long integer)
 {
 	ft_putint_base_fd(integer, 10, 1);
 }
